Code anti-deprecation; monocle v1 => 3, initial cats-effect, graalvm,...
Monocle updated to v3. API undeprectaion.

Cats-effect added into deps, but not used by now.
Started initial GraalVM execises, to make SSR for showcase.

Failed to update to sjs-dom v2: sjs-react also needs to major update,
but diode not yet ready.

+++ src1/client/bill/cart/src/main/scala/io/suggest/bill/cart/c/ILkCartApi.scala
@@ -93,7 +93,7 @@ final class LkCartApiHttpImpl extends ILkCartApi {
   override def unHoldOrder(orderId: Gid_t): Future[MOrderContent] = {
     HttpClient.execute(
       HttpReq.routed(
-        route = routes.controllers.LkBill2.unHoldOrder( orderId ),
+        route = routes.controllers.LkBill2.unHoldOrder( orderId.toDouble ),
       )
     )
       .respAuthFut
